+++ _tooling/publish.sh
@@ -9,14 +9,18 @@ for PKG in {cluster_tools,webknossos}/pyproject.toml; do
     pushd "$PKG" > /dev/null
 
     cp pyproject.toml pyproject.toml.bak
-    PKG_VERSION="$(dunamai from git)"
+    PKG_VERSION="$(uvx dunamai from git)"
 
     echo "__version__ = '$PKG_VERSION'" > ./"$PKG"/version.py
 
-    poetry version "$PKG_VERSION"
-    # replace all relative path dependencies with the current version:
-    sed -i 's/\(.*\) = .* path \= \"\.\..*/\1 = "'"$PKG_VERSION"'"/g' pyproject.toml
-    poetry publish --build
+    # Update version number in pyproject.toml
+    sed -i 's/version = "0.0.0"/version = "'"${PKG_VERSION}"'"/g' pyproject.toml    
+
+    # replace relative path dependencies (i.e. cluster-tools) with the current version:
+    sed -i 's/"cluster-tools"/"cluster-tools=='"${PKG_VERSION}"'"/g' pyproject.toml    
+    
+    uv build
+    uv publish
 
     # Restore files
     mv pyproject.toml.bak pyproject.toml