+++ cluster_tools/cluster_tools/schedulers/kube.py
@@ -54,7 +54,7 @@ class KubernetesExecutor(ClusterExecutor):
             import kubernetes  # noqa: F401 unused import
         except ModuleNotFoundError:
             logger.error(
-                'The Kubernetes Python package is not installed. cluster_tools does not install this dependency be default. Run `pip install cluster_tools[kubernetes]` or `poetry install --extras "kubernetes"` to install Kubernetes support.'
+                'The Kubernetes Python package is not installed. cluster_tools does not install this dependency be default. Run `pip install cluster_tools[kubernetes]` or `uv sync --extra "kubernetes"` to install Kubernetes support.'
             )
             exit()