+++ cluster_tools/test.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
 set -eEuo pipefail
 
-echo "test.sh is not available for cluster_tools yet"
+  cd tests
+  PYTEST_EXECUTORS=multiprocessing,sequential,test_pickling,debug_sequential uv run --frozen python -m pytest -sv test_all.py test_multiprocessing.py
+
+  echo "Tests for the kubernetes, dask and SLURM executors are only run in the CI"